refactor(test)!: inline make_credential locally, drop credential store::test_helpers (test-util eradication, step 1)
- Delete `#[cfg(any(test, feature = "test-util"))] pub mod test_helpers` from
  `nebula-credential::store`; no consumers remain.
- Add `#[cfg(all(test, feature = "test-util"))] pub(crate) mod test_support::make_credential`
  to `nebula-storage::credential::mod` for src-level unit tests; gated to match
  call sites so dead-code lint never fires without test-util.
- Create `crates/storage/tests/common/mod.rs` shared integration-test helper
  with the same `make_credential` builder for `credential_store_conformance`
  and `credential_audit_durable`.
- Repoint all four consumers in `nebula-storage` src tests
  (`layer/encryption`, `layer/cache`, `key_provider` ×2) to
  `crate::credential::test_support::make_credential`.
- `nebula-tenancy` already had a local `make_credential`; no change needed.
