+++ crates/storage/src/credential/layer/cache.rs
@@ -191,7 +191,9 @@ impl<S: CredentialStore> CredentialStore for CacheLayer<S> {
 // `--all-targets` (no features) skip this block.
 #[cfg(all(test, feature = "test-util", feature = "sqlite"))]
 mod tests {
-    use nebula_credential::{PutMode, StoreError, store::test_helpers::make_credential};
+    use nebula_credential::{PutMode, StoreError};
+
+    use crate::credential::test_support::make_credential;
 
     use super::{super::super::sqlite::SqliteCredentialStore, *};
