+++ crates/storage/src/credential/layer/encryption.rs
@@ -250,14 +250,13 @@ impl<S> EncryptionLayer<S> {
     }
 }
 
-// Tests require `nebula-credential/test-util` to reach `test_helpers` and
-// `StaticKeyProvider`. Storage's own `test-util` feature forwards that.
+// `StaticKeyProvider` requires `nebula-credential/test-util`, forwarded by
+// storage's own `test-util` feature.
 #[cfg(all(test, feature = "test-util", feature = "sqlite"))]
 mod tests {
-    use nebula_credential::{
-        AuthStyle, PutMode, SecretString, credentials::oauth2::OAuth2State,
-        store::test_helpers::make_credential,
-    };
+    use nebula_credential::{AuthStyle, PutMode, SecretString, credentials::oauth2::OAuth2State};
+
+    use crate::credential::test_support::make_credential;
     use nebula_crypto::encrypt_with_key_id;
 
     use super::{
